main update 1.3

+++ 1.3/main.cpp
@@ -5,14 +5,14 @@
 #include <fstream>
 #include <stdexcept>
 
-void print_matrix(const std::vector<long double> &x) {
+void print_matrix(const std::vector<double> &x) {
     for (auto elem : x) {
         std::cout << std::fixed << std::setprecision(10) << elem << ' ';
     }
     std::cout << '\n';
 }
 
-void print_matrix(const std::vector<std::vector<long double>> &matrix) {
+void print_matrix(const std::vector<std::vector<double>> &matrix) {
     for (int i = 0; i < matrix.size(); ++i) {
         for (int j = 0; j < matrix[0].size(); ++j) {
             std::cout << std::fixed << std::setprecision(10) << matrix[i][j] << ' ';
@@ -22,8 +22,8 @@ void print_matrix(const std::vector<std::vector<long double>> &matrix) {
 }
 
 // * функция подсчета нормы разности двух векторов
-long double calculate_norm(const std::vector<long double> &x, const std::vector<long double> &prev_x) {
-    long double norm = 0.0;
+double calculate_norm(const std::vector<double> &x, const std::vector<double> &prev_x) {
+    double norm = 0.0;
     for (size_t i = 0; i < x.size(); ++i) {
         norm = std::max(norm, fabs(x[i] - prev_x[i]));
     }
@@ -31,16 +31,16 @@ long double calculate_norm(const std::vector<long double> &x, const std::vector<
 }
 
 // * метод простых итераций
-std::vector<long double> simple_iterations_algorithm(const std::vector<std::vector<long double>> &matrix, const std::vector<long double> &b, long double EPS, int &iter) {
+std::vector<double> simple_iterations_algorithm(const std::vector<std::vector<double>> &matrix, const std::vector<double> &b, double EPS, int &iter) {
     size_t n = matrix.size();
     // * начальное приближение
-    std::vector<long double> x(n, 0.0);
-    std::vector<long double> prev_x(n);
+    std::vector<double> x(n, 0.0);
+    std::vector<double> prev_x(n);
 
     do {
         prev_x = x;
         for (size_t i = 0; i < n; ++i) {
-            long double sum = 0.0;
+            double sum = 0.0;
             for (size_t j = 0; j < n; ++j) {
                 if (i != j) {
                     sum += matrix[i][j] * prev_x[j];
@@ -54,16 +54,16 @@ std::vector<long double> simple_iterations_algorithm(const std::vector<std::vect
     return x;
 }
 
-std::vector<long double> seidel_algorithm(const std::vector<std::vector<long double>> &matrix, const std::vector<long double> &b, long double EPS, int &iter) {
+std::vector<double> seidel_algorithm(const std::vector<std::vector<double>> &matrix, const std::vector<double> &b, double EPS, int &iter) {
     size_t n = matrix.size();
 
     // * начальное приближение
-    std::vector<long double> x(n, 0.0);
+    std::vector<double> x(n, 0.0);
 
     do {
-        std::vector<long double> prev_x = x;
+        std::vector<double> prev_x = x;
         for (size_t i = 0; i < n; ++i) {
-            long double sum = 0.0;
+            double sum = 0.0;
             for (size_t j = 0; j < i; ++j) {
                 sum += matrix[i][j] * x[j];
             }
@@ -136,7 +136,7 @@ int main() {
     int n;
     fin >> n;
 
-    std::vector<std::vector<long double>> matrix(n, std::vector<long double>(n, 0.0));
+    std::vector<std::vector<double>> matrix(n, std::vector<double>(n, 0.0));
 
     for (int i = 0; i < n; ++i) {
         for (int j = 0; j < n; ++j) {
@@ -144,29 +144,29 @@ int main() {
         }
     }
 
-    std::vector<long double> b(n, 0.0);
+    std::vector<double> b(n, 0.0);
     for (int i = 0; i < n; ++i) {
         fin >> b[i];
     }
 
-    long double EPS = 1e-6;
+    double EPS = 1e-6;
 
     fin >> EPS;
 
     int simple_iter = 0;
     int seidel_iter = 0;
 
-    std::vector<long double> simple_iter_result = simple_iterations_algorithm(matrix, b, EPS, simple_iter);
+    std::vector<double> simple_iter_result = simple_iterations_algorithm(matrix, b, EPS, simple_iter);
     std::cout << "> Решение методом простых итераций:" << '\n';
     print_matrix(simple_iter_result);
     std::cout << "> Количество итераций: " << simple_iter << '\n';
 
-    std::vector<long double> gauss_seidel_result = seidel_algorithm(matrix, b, EPS, seidel_iter);
+    std::vector<double> gauss_seidel_result = seidel_algorithm(matrix, b, EPS, seidel_iter);
     std::cout << "\n> Решение методом Зейделя:" << '\n';
     print_matrix(gauss_seidel_result);
     std::cout << "> Количество итераций: " << seidel_iter << '\n';
 
-    std::vector<std::vector<long double>> check_simple_iter;
+    std::vector<std::vector<double>> check_simple_iter;
     try {
         check_simple_iter = multiply_matrices(matrix, simple_iter_result);
     } catch(std::exception &e) {
@@ -176,7 +176,7 @@ int main() {
     std::cout << "\n> Проверка решения методом простых итераций:\n";
     print_matrix(check_simple_iter);
 
-    std::vector<std::vector<long double>> check_seidel;
+    std::vector<std::vector<double>> check_seidel;
     try {
         check_seidel = multiply_matrices(matrix, gauss_seidel_result);
     } catch(std::exception &e) {
